requirements.txt
anthropic>=0.40.0
pymupdf>=1.23.0
Pillow>=10.0.0
streamlit>=1.35.0
pytest>=7.4.0
